refactor : 메인페이지 리팩토링 및 레이아웃 수정 (#466)
* refactor : 3개 편지를 불러오는 데이터 페징 로직 훅 분리

* refactor : 푸시알림 훅 분리

* design : 메인페이지 레이아웃 분리 및 슬라이드 배너 하단 고정

* chore : 수정

* refactor : 그라디언트와 물결 컴포넌트로 분리

* design : 홈화면에서 내비게이션바뒤로 배경이 보이던 문제 해결
