+++ src/components/Common/NavigationBar/NavigationBar.tsx
@@ -18,7 +18,7 @@ export const NavigationBar = () => {
     ];
 
     return (
-        <nav className="flex justify-around bg-white border-y p-3 max-w-[473px] min-w-[375px] w-full z-[9999]">
+        <nav className="flex justify-around bg-white border-y p-3 max-w-[475px] min-w-[375px] w-full z-[9999]">
             {navItems.map((item) => (
                 <div className="flex justify-center flex-1" key={item.id}>
                     <NavigationItem
