+++ src/pages/Home/HomePage.tsx
@@ -5,130 +5,37 @@ import { BottomSheetContent } from '@/components/HomePage/BottomSheet/BottomShee
 import { LetterContainer } from '@/components/HomePage/LetterContainer/LetterContainer';
 import { WelcomeMessageContainer } from '@/components/HomePage/WelcomeMessageContainer/WelcomeMessageContainer';
 import { useUserStore } from '@/stores/useUserStore';
-import { useEffect, useState } from 'react';
-import { useGetRecommendLetter } from '@/hooks/useGetRecommendLetter';
-import { useGetRecentRelyLetter } from '@/hooks/useGetRecentRelyLetter';
+import { useState } from 'react';
 import { TopButtonContainer } from '@/components/HomePage/TopButtonContainer/TopButtonContainer';
-import { getToken, firebaseMessaging, onMessage } from '@/util/firebase';
-import { postToken } from '@/service/nofication/postToken';
-import { useToastStore } from '@/hooks';
 import { Loading } from '@/components/Common/Loading/Loading';
 import { ToggleVariant } from '@/constants/toggleVariant';
-
-export type ReplyLetter = {
-    type: 'MAP' | 'KEYWORD';
-    labelUrl: string;
-    letterId: number;
-};
-
-export type RecommendLetter = {
-    letterId: number;
-    title: string;
-    label: string;
-};
+import { useGetThreeLetterData } from '@/hooks/useGetThreeLetterData';
+import { ErrorPage } from '../ErrorPage';
+import { usePushNotifications } from '@/hooks/usePushNotifications';
+import { Margin } from '@/components/Common/Margin/Margin';
 
 const HomePage = () => {
+    usePushNotifications(); // 푸시 알림 훅
+
     const { user } = useUserStore();
-    const { addToast } = useToastStore();
-
-    const isToken = localStorage.getItem('isToken');
-
-    useEffect(() => {
-        if (isToken === 'true') {
-            return;
-        }
-        const handlePushNotifications = async () => {
-            try {
-                const permission = await Notification.requestPermission();
-
-                if (permission === 'granted') {
-                    const token = await getToken(firebaseMessaging, {
-                        vapidKey: import.meta.env.VITE_FCM_VAPID_KEY
-                    });
-
-                    try {
-                        const data = await postToken({ token });
-                        addToast(`${data.message}`, 'success');
-                        localStorage.setItem('isToken', 'true');
-                    } catch (error) {
-                        console.error(error);
-                    }
-                } else {
-                    console.log('Notification permission denied');
-                }
-            } catch (error) {
-                console.error('FCM Error : ', error);
-            }
-        };
-
-        // 푸시 알림 설정 및 수신 처리
-        handlePushNotifications();
-
-        // 푸시 알림 메시지 수신 시 처리
-        const unsubscribe = onMessage(firebaseMessaging, (payload) => {
-            if (payload.notification) {
-                alert(payload.notification.body);
-            }
-        });
-
-        // 컴포넌트가 언마운트 될 때 리스너 해제
-        return () => {
-            unsubscribe();
-        };
-    }, []); // 빈 배열을 의존성으로 사용해 처음 마운트될 때만 실행되게 함
-
     const [open, setOpen] = useState(false);
     const [toggle, setToggle] = useState(true);
-    const [letters, setLetters] = useState<ReplyLetter[] | RecommendLetter[]>(
-        []
-    );
 
-    function onDismiss() {
-        setOpen(false);
+    const { letters, isLoading, isError } = useGetThreeLetterData(toggle);
+
+    if (isLoading) {
+        return <Loading />;
     }
 
-    const {
-        // data: recommendedLetterData,
-        refetch: refetchRecommendedLetters,
-        isLoading: isRecommendedLetterLoading,
-        isError: isRecommendedLetterError
-    } = useGetRecommendLetter();
-
-    const {
-        // data: recentRelyLetterData,
-        refetch: refetchRecentRelyLetters,
-        isLoading: isRecentRelyLetterLoading,
-        isError: isRecentRelyLetterError
-    } = useGetRecentRelyLetter();
-
-    useEffect(() => {
-        if (toggle) {
-            refetchRecommendedLetters().then((response) => {
-                setLetters(response?.data?.result || []);
-            });
-        } else {
-            refetchRecentRelyLetters().then((response) => {
-                setLetters(response?.data?.result || []);
-            });
-        }
-    }, [toggle, refetchRecommendedLetters, refetchRecentRelyLetters]);
-
-    if (isRecommendedLetterLoading || isRecentRelyLetterLoading) {
-        return (
-            <div className="flex flex-1 w-full h-full">
-                <Loading />
-            </div>
-        );
-    }
-
-    if (isRecommendedLetterError || isRecentRelyLetterError) {
-        return <p>데이터를 불러오는 중 오류가 발생했습니다.</p>;
+    if (isError) {
+        return <ErrorPage />;
     }
 
     return (
-        <>
-            <TopButtonContainer />
-            <div className="flex flex-col gap-5">
+        <div className="relative h-full w-full flex flex-col justify-between gap-5 ">
+            <div className="z-0">
+                <TopButtonContainer />
+
                 <Toggle
                     isChecked={!toggle}
                     onToggle={() => {
@@ -139,13 +46,13 @@ const HomePage = () => {
                     variant={ToggleVariant.Main}
                 />
 
-                <div>
-                    <WelcomeMessageContainer
-                        nickname={user?.nickname}
-                        newLetter={letters.length > 0}
-                    />
-                    <LetterContainer letters={letters} />
-                </div>
+                <WelcomeMessageContainer
+                    nickname={user?.nickname}
+                    newLetter={letters.length > 0}
+                />
+
+                <LetterContainer letters={letters} />
+
                 <div className="flex justify-center px-20">
                     <button
                         onClick={() => {
@@ -156,25 +63,26 @@ const HomePage = () => {
                         키워드 설정
                     </button>
                 </div>
-
-                <div className="mx-[-20px]">
-                    <BannerContainer />
-                </div>
-
-                <SliderMenuContainer
-                    open={open}
-                    onDismiss={onDismiss}
-                    snapPoints={() => [window.innerHeight * 0.95]}
-                >
-                    <BottomSheetContent
-                        onClick={() => {
-                            setOpen(false);
-                        }}
-                        nickname={user?.nickname}
-                    />
-                </SliderMenuContainer>
             </div>
-        </>
+
+            <div className="relative -mx-5">
+                <BannerContainer />
+                <Margin bottom={20} />
+            </div>
+
+            <SliderMenuContainer
+                open={open}
+                onDismiss={() => setOpen(false)}
+                snapPoints={() => [window.innerHeight * 0.95]}
+            >
+                <BottomSheetContent
+                    onClick={() => {
+                        setOpen(false);
+                    }}
+                    nickname={user?.nickname}
+                />
+            </SliderMenuContainer>
+        </div>
     );
 };
